Move remaining src/ingestion under src/shared
