+++ benchmarks/postgres/ingestion.benchmark.ts
@@ -4,8 +4,8 @@ import os from 'os'
 import { performance } from 'perf_hooks'
 
 import { IEvent } from '../../src/idl/protos'
-import { EventsProcessor } from '../../src/ingestion/process-event'
 import { createServer } from '../../src/server'
+import { EventsProcessor } from '../../src/shared/ingestion/process-event'
 import { LogLevel, PluginsServer, SessionRecordingEvent, Team } from '../../src/types'
 import { UUIDT } from '../../src/utils'
 import { getFirstTeam, resetTestDatabase } from '../../tests/helpers/sql'
