+++ src/config.ts
@@ -1,6 +1,6 @@
 import os from 'os'
 
-import { KAFKA_EVENTS_PLUGIN_INGESTION } from './ingestion/topics'
+import { KAFKA_EVENTS_PLUGIN_INGESTION } from './shared/ingestion/topics'
 import { LogLevel, PluginsServerConfig } from './types'
 
 export const defaultConfig = overrideWithEnv(getDefaultConfig())
