+++ src/worker/ingestion/process-event.ts
@@ -9,9 +9,14 @@ import { nodePostHog } from 'posthog-js-lite/dist/src/targets/node'
 
 import { DB } from '../../db'
 import { Event as EventProto, IEvent } from '../../idl/protos'
-import { KAFKA_EVENTS, KAFKA_SESSION_RECORDING_EVENTS } from '../../ingestion/topics'
-import { elementsToString, personInitialAndUTMProperties, sanitizeEventName, timeoutGuard } from '../../ingestion/utils'
 import Client from '../../shared/celery/client'
+import { KAFKA_EVENTS, KAFKA_SESSION_RECORDING_EVENTS } from '../../shared/ingestion/topics'
+import {
+    elementsToString,
+    personInitialAndUTMProperties,
+    sanitizeEventName,
+    timeoutGuard,
+} from '../../shared/ingestion/utils'
 import { status } from '../../status'
 import {
     CohortPeople,
